refactor:エラー内容の修正とファイル構成の修正。#172
