+++ frontend/src/features/management/components/groupSelector/groupSelector.module.css
