+++ frontend/src/features/management/components/groupSelector/groupSelector.tsx
