+++ frontend/src/features/management/style/editNotice.module.css
