+++ frontend/src/features/management/style/noticeCard.module.css
