+++ frontend/src/features/management/types/notice.ts
